A simple settings screen with just a logout button
